Fixed demo data generation for living doc and dashboard data: - wrong target path for dashboard data - ensure clean of output directory done in right place (not after new data was already generated)
