+++ scenarioo-docu-generation-example/src/test/java/org/scenarioo/example/livingDoc/PizzaShopExampleLivingDocGeneratorTest.java
@@ -1,6 +1,7 @@
 package org.scenarioo.example.livingDoc;
 
 import org.junit.Test;
+import org.scenarioo.example.util.BuildOutputDirResource;
 import org.scenarioo.model.docu.entities.Branch;
 import org.scenarioo.model.docu.entities.Build;
 
@@ -17,7 +18,7 @@ public class PizzaShopExampleLivingDocGeneratorTest {
 		// Config of generator:
 		String branchName = "PizzaaaS - Living Doc Example";
 		String buildName = "build-1";
-		File targetDir = new File("build/scenarioDocuExample");
+		File targetDir = BuildOutputDirResource.DOCU_BUILD_DIRECTORY;
 		File docFilesSourceDir = getResourceFile("example/pizza-shop-example");
 		LivingDocGenerator generator = new LivingDocGenerator(targetDir, branchName, buildName);
