+++ scenarioo-docu-generation-example/src/test/java/org/scenarioo/uitest/example/infrastructure/MultipleBuildsRule.java
@@ -12,6 +12,7 @@ import org.junit.rules.TestRule;
 import org.junit.runner.Description;
 import org.junit.runners.model.Statement;
 import org.scenarioo.api.ScenarioDocuWriter;
+import org.scenarioo.example.util.BuildOutputDirResource;
 import org.scenarioo.model.docu.entities.Build;
 
 public class MultipleBuildsRule implements TestRule {
@@ -19,9 +20,9 @@ public class MultipleBuildsRule implements TestRule {
 	private static final Logger LOGGER = Logger.getLogger(MultipleBuildsRule.class);
 
 	private static BuildRun currentBuildRun;
-	
+
 	private final SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
-	
+
 	@Override
 	public Statement apply(final Statement base, final Description description) {
 		return new Statement() {
@@ -38,9 +39,10 @@ public class MultipleBuildsRule implements TestRule {
 					e.printStackTrace();
 				}
 			}
-			
+
 			public void write_build_description() {
-				ScenarioDocuWriter docuWriter = new ScenarioDocuWriter(DOCU_BUILD_DIRECTORY, getCurrentBranchName(), getCurrentBuildName());
+				ScenarioDocuWriter docuWriter = new ScenarioDocuWriter(BuildOutputDirResource.DOCU_BUILD_DIRECTORY,
+					getCurrentBranchName(), getCurrentBuildName());
 				Build build = new Build();
 				build.setName(getCurrentBuildName());
 				build.setDate(getDate());
@@ -59,18 +61,18 @@ public class MultipleBuildsRule implements TestRule {
 			}
 		};
 	}
-	
+
 	public static BuildRun getCurrentBuildRun() {
 		return currentBuildRun;
 	}
-	
+
 	public static String getCurrentBranchName() {
 		return currentBuildRun.getBranchName();
 	}
-	
+
 	public static String getCurrentBuildName() {
 		return currentBuildRun.getDate();
 	}
 
-	
+
 }
